修复PC打包BUG

+++ game/script.rpy
@@ -65,10 +65,11 @@ label main_menu:
 label quit:
     python:
         import os
-        cache_path = config.gamedir + f"\\audio\\ai_audio"
-        for file in os.listdir(cache_path):
-            file = f"{cache_path}\\{file}"
-            os.remove(file)
+        cache_path = os.path.join(config.gamedir, "/audio/ai_audio")
+        if os.path.exists(cache_path):
+            for file in os.listdir(cache_path):
+                file = f"{cache_path}/{file}"
+                os.remove(file)
     return
 
 
@@ -119,12 +120,12 @@ label start:
         me "[user_input!q]"
 
         while chat.waiting:
-            e "......"
+            pause
 
         if chat.msg:
             $ msg = chat.msg[-1]["content"]
         else:
-            call conn_error(err_info=chat.error)
+            call conn_error(err_info=chat.error) from _call_conn_error
 
         $ words = chat.parse_words(msg)
 
@@ -138,7 +139,7 @@ label start:
             if persistent.voice_source == "ai":
                 $ renpy.invoke_in_thread(chat.get_ai_voice, w)
                 while chat.waiting:
-                    e "......" 
+                    pause
                 $ vc = chat.current_vc if chat.current_vc else chat.get_chat_state(w, Chat.CHAT_VOICE)
             else:
                 $ vc = chat.get_chat_state(w, Chat.CHAT_VOICE)
